+++ CHANGELOG.md
@@ -3,6 +3,81 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.1](https://github.com/baloise/ui-library-next/compare/v1.13.0...v1.13.1) (2020-11-09)
+
+
+### Bug Fixes
+
+* **select:** remove loading from readonly ([a351a27](https://github.com/baloise/ui-library-next/commit/a351a27de49bbc62bffda6907805c4ca505a8386))
+
+
+
+
+
+# [1.13.0](https://github.com/baloise/ui-library-next/compare/v1.12.3...v1.13.0) (2020-11-09)
+
+
+### Features
+
+* **select:** new data prop for better rendering custom layouts ([9916465](https://github.com/baloise/ui-library-next/commit/9916465db43ce0a2871af7add9ee1a74c6ef93c1))
+* **select:** new data prop for better rendering custom layouts ([869f28e](https://github.com/baloise/ui-library-next/commit/869f28e94db691d1ebe287fa7ab178d4ea92ae99))
+
+
+
+
+
+## [1.12.3](https://github.com/baloise/ui-library-next/compare/v1.12.2...v1.12.3) (2020-11-09)
+
+
+### Bug Fixes
+
+* **select:** improve open and close of the dropdown ([c57e5ca](https://github.com/baloise/ui-library-next/commit/c57e5ca995eb4872f96c0b8494bc48d384f257c0))
+
+
+
+
+
+## [1.12.2](https://github.com/baloise/ui-library-next/compare/v1.12.1...v1.12.2) (2020-11-09)
+
+
+### Bug Fixes
+
+* **vue:** array binding ([93a8a4b](https://github.com/baloise/ui-library-next/commit/93a8a4b6deeedd85d8d31d25b45ee418806ca0ec))
+
+
+
+
+
+## [1.12.1](https://github.com/baloise/ui-library-next/compare/v1.12.0...v1.12.1) (2020-11-09)
+
+
+### Bug Fixes
+
+* **vue:** array binding ([d879bb0](https://github.com/baloise/ui-library-next/commit/d879bb0c237da3b1abaf2e5d2e99b4a5a68284d4))
+
+
+
+
+
+# [1.12.0](https://github.com/baloise/ui-library-next/compare/v1.11.1...v1.12.0) (2020-11-09)
+
+
+### Features
+
+* improve components ([17cba42](https://github.com/baloise/ui-library-next/commit/17cba4246815b2aa92ea83c42bd29ba2b0a6ee77))
+
+
+
+
+
+## [1.11.1](https://github.com/baloise/ui-library-next/compare/v1.11.0...v1.11.1) (2020-11-09)
+
+**Note:** Version bump only for package root
+
+
+
+
+
 # [1.11.0](https://github.com/baloise/ui-library-next/compare/v1.10.0...v1.11.0) (2020-11-08)