+++ packages/testing/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.1](https://github.com/baloise/ui-library-next/compare/v1.13.0...v1.13.1) (2020-11-09)
+
+**Note:** Version bump only for package @baloise/ui-library-next-testing
+
+
+
+
+
+# [1.13.0](https://github.com/baloise/ui-library-next/compare/v1.12.3...v1.13.0) (2020-11-09)
+
+**Note:** Version bump only for package @baloise/ui-library-next-testing
+
+
+
+
+
+## [1.12.3](https://github.com/baloise/ui-library-next/compare/v1.12.2...v1.12.3) (2020-11-09)
+
+**Note:** Version bump only for package @baloise/ui-library-next-testing
+
+
+
+
+
+## [1.12.2](https://github.com/baloise/ui-library-next/compare/v1.12.1...v1.12.2) (2020-11-09)
+
+**Note:** Version bump only for package @baloise/ui-library-next-testing
+
+
+
+
+
+## [1.12.1](https://github.com/baloise/ui-library-next/compare/v1.12.0...v1.12.1) (2020-11-09)
+
+**Note:** Version bump only for package @baloise/ui-library-next-testing
+
+
+
+
+
+# [1.12.0](https://github.com/baloise/ui-library-next/compare/v1.11.1...v1.12.0) (2020-11-09)
+
+**Note:** Version bump only for package @baloise/ui-library-next-testing
+
+
+
+
+
+## [1.11.1](https://github.com/baloise/ui-library-next/compare/v1.11.0...v1.11.1) (2020-11-09)
+
+**Note:** Version bump only for package @baloise/ui-library-next-testing
+
+
+
+
+
 # [1.11.0](https://github.com/baloise/ui-library-next/compare/v1.10.0...v1.11.0) (2020-11-08)
 
 **Note:** Version bump only for package @baloise/ui-library-next-testing